requirements.txt
python-telegram-bot
requests
pymediainfo
pytube
bs4